# Conflicts:
#	ndtp-user/src/main/webapp/css/ko/user-style.css
#	ndtp-user/src/main/webapp/js/ko/map-controll.js

+++ ndtp-admin/src/main/java/ndtp/service/impl/LayerServiceImpl.java
@@ -80,6 +80,46 @@ public class LayerServiceImpl implements LayerService {
     public List<Layer> getListLayer(Layer layer) {
         return layerMapper.getListLayer(layer);
     }
+    
+    /**
+     * geoserver layer 목록 조회 
+     */
+    @Transactional(readOnly=true)
+    public String getListGeoserverLayer(GeoPolicy geoPolicy) {
+    	String geoserverLayerJson = null;
+    	try {
+			RestTemplate restTemplate = new RestTemplate();
+			
+			HttpHeaders headers = new HttpHeaders();
+			// 클라이언트가 서버에 어떤 형식(MediaType)으로 달라는 요청을 할 수 있는데 이게 Accpet 헤더를 뜻함.
+			List<MediaType> acceptList = new ArrayList<>();
+			acceptList.add(MediaType.ALL);
+			headers.setAccept(acceptList);
+			// 클라이언트가 request에 실어 보내는 데이타(body)의 형식(MediaType)를 표현
+			headers.setContentType(MediaType.TEXT_XML);
+			// geoserver basic 암호화 아이디:비밀번호 를 base64로 encoding 
+			headers.add("Authorization", "Basic " + Base64.getEncoder().
+					encodeToString((geoPolicy.getGeoserverUser() + ":" + geoPolicy.getGeoserverPassword()).getBytes()));
+			
+			List<HttpMessageConverter<?>> messageConverters = new ArrayList<HttpMessageConverter<?>>();
+			//Add the String Message converter
+			messageConverters.add(new StringHttpMessageConverter());
+			//Add the message converters to the restTemplate
+			restTemplate.setMessageConverters(messageConverters);
+		    
+			HttpEntity<String> entity = new HttpEntity<>(headers);
+			
+			String url = geoPolicy.getGeoserverDataUrl() + "/rest/workspaces/" + geoPolicy.getGeoserverDataWorkspace()+ "/layers";
+			ResponseEntity<?> response = restTemplate.exchange(url, HttpMethod.GET, entity, String.class);
+			log.info("-------- statusCode = {}, body = {}", response.getStatusCodeValue(), response.getBody());
+			geoserverLayerJson = response.getBody().toString();
+			
+		} catch(Exception e) {
+			e.printStackTrace();
+		}
+    	
+    	return geoserverLayerJson;
+    }
 
     /**
     * layer 정보 취득
@@ -89,6 +129,11 @@ public class LayerServiceImpl implements LayerService {
     @Transactional(readOnly=true)
     public Layer getLayer(Integer layerId) {
         return layerMapper.getLayer(layerId);
+    }
+    
+    @Transactional(readOnly=true)
+    public Boolean isLayerKeyDuplication(String layerKey) {
+    	return layerMapper.isLayerKeyDuplication(layerKey);
     }
 
     /**
@@ -151,6 +196,7 @@ public class LayerServiceImpl implements LayerService {
             layerFileInfoGroupMap.put("shapeEncoding", shapeEncoding);
             layerFileInfoGroupMap.put("layerFileInfoGroupId", layerFileInfoGroupId);
             layerFileInfoGroupMap.put("layerFileInfoGroupIdList", layerFileInfoGroupIdList);
+            layerFileInfoGroupMap.put("layerId", layerId);
             log.info("+++ layerFileInfoGroupMap = {}", layerFileInfoGroupMap);
             layerFileInfoMapper.updateLayerFileInfoGroup(layerFileInfoGroupMap);
         }
@@ -248,7 +294,7 @@ public class LayerServiceImpl implements LayerService {
         }
 
         GeoPolicy geoPolicy = geoPolicyService.getGeoPolicy();
-        String layerSourceCoordinate = geoPolicy.getLayerSourceCoordinate();
+        String layerSourceCoordinate = layer.getCoordinate();
         String layerTargetCoordinate = geoPolicy.getLayerTargetCoordinate();
 //		ShapeFileParser shapeFileParser = new ShapeFileParser();
 //		shapeFileParser.parse(shapeFileName);
@@ -375,7 +421,10 @@ public class LayerServiceImpl implements LayerService {
 			// layer_file_info 히스토리 삭제
 			layerFileInfoMapper.deleteLayerFileInfo(layerId);
 			// 공간정보 테이블 삭제
-			layerMapper.deleteLayerTable(layer.getLayerKey());
+			String layerExists = layerMapper.isLayerExists(layer.getLayerKey());
+			if(layerExists != null) {
+				layerMapper.deleteLayerTable(layer.getLayerKey());
+			}
 		}
 		// 레이어 메타정보 삭제 
 		return layerMapper.deleteLayer(layerId);
